ProjectCode/requirements.txt
Flask
graphviz
plantuml
pydot
PyYAML
Werkzeug
flask
flask-login
flask-wtf
email-validator
bcrypt
pyotp
qrcode
PyJWT
cryptography


